Tweak rewards, domain randomization, add keyboard control

+++ genesis_playground/zbot/zbot_train.py
@@ -1,13 +1,16 @@
 """ ZBot training
 
 Run:
-    python genesis_playground/zbot/zbot_train.py --num_envs 4096
+    python genesis_playground/zbot/zbot_train.py --num_envs 4096 --max_iterations 200 --device mps 
+    (or cuda)
 """
 
 import argparse
 import os
 import pickle
 import shutil
+import wandb
+from datetime import datetime
 
 from zbot_env import ZbotEnv
 from rsl_rl.runners import OnPolicyRunner
@@ -66,6 +69,7 @@ def get_cfgs():
     env_cfg = {
         "num_actions": 10,
         # joint/link names
+        # NOTE: hip roll/yaw flipped between sim & real robot FIXME
         "default_joint_angles": {  # [rad]
             "R_Hip_Pitch": 0.0,
             "L_Hip_Pitch": 0.0,
@@ -90,9 +94,24 @@ def get_cfgs():
             "R_Ankle_Pitch",
             "L_Ankle_Pitch",
         ],
+        # friction
+        "env_friction_range": {
+            "start": [0.9, 1.1],
+            "end": [0.9, 1.1],
+        },
+        # link mass
+        # varying this too much collapses the training
+        "link_mass_multipliers": {
+            "start": [1.0, 1.0],
+            "end": [1.0, 1.0],
+        },
+        # RFI
+        "rfi_scale": 0.1,
         # PD
         "kp": 20.0,
         "kd": 0.5,
+        "kp_multipliers": [0.75, 1.25],
+        "kd_multipliers": [0.75, 1.25],
         # termination
         "termination_if_roll_greater_than": 10,  # degree
         "termination_if_pitch_greater_than": 10,
@@ -104,9 +123,11 @@ def get_cfgs():
         "action_scale": 0.25,
         "simulate_action_latency": True,
         "clip_actions": 100.0,
+        "max_torque": 10.0,
     }
     obs_cfg = {
         "num_obs": 39,
+        # FIXME: IMU mounting orientation is different between sim & real robot
         "obs_scales": {
             "lin_vel": 2.0,
             "ang_vel": 0.25,
@@ -125,36 +146,54 @@ def get_cfgs():
             "base_height": -50.0,
             "action_rate": -0.005,
             "similar_to_default": -0.1,
+            "feet_air_time": 5.0,
         },
     }
     command_cfg = {
         "num_commands": 3,
-        "lin_vel_y_range": [0.0, 0.0], # move faster than above!
-        "lin_vel_x_range": [0.1, 0.4],
-        "ang_vel_range": [0, 0],
+        "lin_vel_y_range": [0.0, 0.0],
+        "lin_vel_x_range": [-0.2, 0.4],
+        "ang_vel_range": [-0.4, 0.4],
     }
 
     return env_cfg, obs_cfg, reward_cfg, command_cfg
 
+class WandbOnPolicyRunner(OnPolicyRunner):
+    def log(self, info):
+        super().log(info)
+        # Log metrics to wandb
+        metrics = {
+            'train/rew_tracking_lin_vel': info['train/rew_tracking_lin_vel'],
+            'train/rew_tracking_ang_vel': info['train/rew_tracking_ang_vel'],
+            'train/rew_lin_vel_z': info['train/rew_lin_vel_z'],
+            'train/rew_base_height': info['train/rew_base_height'],
+            'train/rew_action_rate': info['train/rew_action_rate'],
+            'train/rew_similar_to_default': info['train/rew_similar_to_default'],
+        }
+        wandb.log(metrics)
 
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("-e", "--exp_name", type=str, default="zbot-walking")
     parser.add_argument("-B", "--num_envs", type=int, default=10)
-    parser.add_argument("--max_iterations", type=int, default=3000)
+    parser.add_argument("--max_iterations", type=int, default=300)
     parser.add_argument("--device", type=str, default="mps")
     parser.add_argument("--show_viewer", type=bool, default=False)
+    parser.add_argument("--wandb_entity", type=str, default=None)
+    parser.add_argument("--use_wandb", type=bool, default=False)
+    parser.add_argument("--from_checkpoint", type=bool, default=False)
     args = parser.parse_args()
-
+    
     gs.init(logging_level="warning")
 
     log_dir = f"logs/{args.exp_name}"
     env_cfg, obs_cfg, reward_cfg, command_cfg = get_cfgs()
     train_cfg = get_train_cfg(args.exp_name, args.max_iterations)
 
-    if os.path.exists(log_dir):
-        shutil.rmtree(log_dir)
-    os.makedirs(log_dir, exist_ok=True)
+    if not args.from_checkpoint:
+        if os.path.exists(log_dir):
+            shutil.rmtree(log_dir)
+        os.makedirs(log_dir, exist_ok=True)
 
     env = ZbotEnv(
         num_envs=args.num_envs, 
@@ -166,14 +205,36 @@ def main():
         show_viewer=args.show_viewer,
     )
 
-    runner = OnPolicyRunner(env, train_cfg, log_dir, device=args.device)
-
     pickle.dump(
         [env_cfg, obs_cfg, reward_cfg, command_cfg, train_cfg],
         open(f"{log_dir}/cfgs.pkl", "wb"),
     )
+    
+    if args.use_wandb:
+        run = wandb.init(
+                project=args.exp_name,
+                entity=args.wandb_entity,
+                name=f"{args.exp_name}_{datetime.now().strftime('%Y%m%d_%H%M%S')}",
+                config={
+                    "num_envs": args.num_envs,
+                    "max_iterations": args.max_iterations,
+                    "device": args.device,
+                    "env_cfg": env_cfg,
+                    "obs_cfg": obs_cfg,
+                    "reward_cfg": reward_cfg,
+                    "command_cfg": command_cfg,
+                    "train_cfg": train_cfg,
+                }
+        )
+        runner = WandbOnPolicyRunner(env, train_cfg, log_dir, device=args.device)
+    else:
+        runner = OnPolicyRunner(env, train_cfg, log_dir, device=args.device)
 
-    runner.learn(num_learning_iterations=args.max_iterations, init_at_random_ep_len=True)
+    try:
+        runner.learn(num_learning_iterations=args.max_iterations, init_at_random_ep_len=True)
+    finally:
+        if args.use_wandb:
+            wandb.finish()
 
 
 if __name__ == "__main__":
